+++ vue/vue_demo2/src/components/admin/plane/show.vue
@@ -25,7 +25,7 @@
 									<i slot="prefix" class="el-input__icon el-icon-search"></i>
 								</el-input>
 							</el-col>
-							<el-col :span="2">
+							<el-button-group>
 								<el-popover placement="top" width="160" v-model="allDel">
 									<p>确定一次删除所选内容吗？</p>
 									<div style="text-align: right; margin: 0">
@@ -36,8 +36,7 @@
 										<i class="el-icon-delete"></i>
 									</el-button>
 								</el-popover>
-							</el-col>
-							<el-col :span="2">
+							
 								<el-popover placement="top" width="200">
 									<p>导出excel</p>
 									<div style="text-align: right; margin: 0">
@@ -48,21 +47,22 @@
 										<i class="el-icon-document-add"></i>
 									</el-button>
 								</el-popover>
-							</el-col>
-							<el-col :span="2">
+							
 								<el-button size="mini">
 									<i class="el-icon-printer"></i>
 								</el-button>
-							</el-col>
+							</el-button-group>
 						</el-row>
 					</template>
 					<template slot-scope="scope">
+						<el-button-group>
 						<el-button size="mini" @click="handleEdit(scope.$index, scope.row)">
 							<i class="el-icon-edit"></i>
 						</el-button>
 						<el-button type="danger" size="mini" @click="openDel(scope.$index, scope.row)">
 							<i class="el-icon-delete"></i>
 						</el-button>
+						</el-button-group>
 					</template>
 				</el-table-column>
 			</el-table>
@@ -187,6 +187,10 @@
 				this.del = false;
 			},
 			deleteAll() {
+				if(typeof this.multipleSelection=='undefined'){
+					this.$message.error("请选择要删除的数据");
+					return false;
+				}
 				this.delData = this.multipleSelection
 				this.delete();
 				this.allDel = false;
@@ -279,6 +283,10 @@
 					})
 			},
 			exportSelect() {
+				if(typeof this.multipleSelection=='undefined'){
+					this.$message.error("请选择要导出的数据");
+					return false;
+				}
 				this.$axios
 					.post('/admin/export/planes', {
 						'companies': this.multipleSelection
